refactor_check_py
